refactor(cli): move helpers out of command directory
