+++ packages/cdktf-cli/bin/cmds/destroy.ts
@@ -2,8 +2,8 @@ import * as yargs from "yargs";
 import React from "react";
 import { Destroy } from "./ui/destroy";
 import { readConfigSync } from "../../lib/config";
-import { renderInk } from "./render-ink";
-import { displayVersionMessage } from "./version-check";
+import { renderInk } from "./helper/render-ink";
+import { displayVersionMessage } from "./helper/version-check";
 import { throwIfNotProjectDirectory } from "./helper/check-directory";
 
 const config = readConfigSync();
