+++ packages/cdktf-cli/bin/cmds/helper/render-ink.tsx
@@ -1,8 +1,8 @@
 import { render } from "ink";
 import React from "react";
 import { exit } from "process";
-import { App } from "./ui/app";
-import { terraformCheck } from "./terraform-check";
+import { App } from "../ui/app";
+import { terraformCheck } from "../helper/terraform-check";
 
 export const renderInk = async (component: React.ReactElement) => {
   await terraformCheck();
