+++ packages/cdktf-cli/bin/cmds/helper/version-check.ts
@@ -15,7 +15,7 @@ export const DISPLAY_VERSION = `${versionNumber()}`;
 
 export function versionNumber(): string {
   // eslint-disable-next-line @typescript-eslint/no-require-imports
-  return require("../../package.json").version.replace(/\+[0-9a-f]+$/, "");
+  return require("../../../package.json").version.replace(/\+[0-9a-f]+$/, "");
 }
 
 export class VersionCheckTTL {
