+++ packages/cdktf-cli/lib/errors.ts
@@ -1,5 +1,5 @@
 import { ReportParams, ReportRequest } from "./checkpoint";
-import { versionNumber } from "../bin/cmds/version-check";
+import { versionNumber } from "../bin/cmds/helper/version-check";
 
 // Errors that will emit telemetry events
 async function report(command: string, payload: Record<string, any>) {
